refacto(read): move `FrmbMenuItem` to new module `menu`
